Move generate_uuid method to sortinghat.utils
This commit moves the generate_uuid method from sortinghat.core.utils
to sortinghat.utils to be reusable from outside the package.

Rename sortinghat.core.utils to sortinghat.core.aux.

Update the tests accordingly.
