+++ sortinghat/core/api.py
@@ -20,7 +20,6 @@
 #     Miguel Ángel Fernández <mafesan@bitergia.com>
 #
 
-import hashlib
 import logging
 
 import django.db.transaction
@@ -57,72 +56,11 @@ from .errors import (InvalidValueError,
                      EqualIndividualError)
 from .log import TransactionsLog
 from .models import Identity, MIN_PERIOD_DATE, MAX_PERIOD_DATE
-from .utils import unaccent_string, merge_datetime_ranges
+from .aux import merge_datetime_ranges
+from ..utils import generate_uuid
 
 
 logger = logging.getLogger(__name__)
-
-
-def generate_uuid(source, email=None, name=None, username=None):
-    """Generate a UUID related to identity data.
-
-    Based on the input data, the function will return the UUID
-    associated to an identity. On this version, the UUID will
-    be the SHA1 of `source:email:name:username` string.
-
-    This string is case insensitive, which means same values
-    for the input parameters in upper or lower case will produce
-    the same UUID.
-
-    The value of `name` will converted to its unaccent form which
-    means same values with accent or unaccent chars (i.e 'ö and o')
-    will generate the same UUID.
-
-    For instance, these combinations will produce the same UUID:
-        ('scm', 'jsmith@example.com', 'John Smith', 'jsmith'),
-        ('scm', 'jsmith@example,com', 'Jöhn Smith', 'jsmith'),
-        ('scm', 'jsmith@example.com', 'John Smith', 'JSMITH'),
-        ('scm', 'jsmith@example.com', 'john Smith', 'jsmith')
-
-    :param source: data source
-    :param email: email of the identity
-    :param name: full name of the identity
-    :param username: user name used by the identity
-
-    :returns: a universal unique identifier for Sorting Hat
-
-    :raises ValueError: when source is `None` or empty; each one
-        of the parameters is `None`; or the parameters are empty.
-    """
-    def to_str(value, unaccent=False):
-        s = str(value)
-        if unaccent:
-            return unaccent_string(s)
-        else:
-            return s
-
-    if source is None:
-        raise ValueError("'source' cannot be None")
-    if source == '':
-        raise ValueError("'source' cannot be an empty string")
-    if not (email or name or username):
-        raise ValueError("identity data cannot be empty")
-
-    s = ':'.join((to_str(source),
-                  to_str(email),
-                  to_str(name, unaccent=True),
-                  to_str(username))).lower()
-    s = s.encode('UTF-8', errors="surrogateescape")
-
-    sha1 = hashlib.sha1(s)
-    uuid = sha1.hexdigest()
-
-    logger.debug(
-        f"UUID '{uuid}' generated; "
-        f"name='{name}' email='{email}' username='{username}' source='{source}';"
-    )
-
-    return uuid
 
 
 @django.db.transaction.atomic
