+++ tests/test_utils.py
@@ -19,364 +19,14 @@
 #     Santiago Dueñas <sduenas@bitergia.com>
 #
 
-import datetime
-
-from dateutil.tz import UTC
 
 from django.test import TestCase
 
-from sortinghat.core.utils import merge_datetime_ranges, unaccent_string, validate_field
+from sortinghat.utils import unaccent_string, generate_uuid
 
-
-CANT_COMPARE_DATES_ERROR = "can't compare offset-naive and offset-aware datetimes"
-DATE_OUT_OF_BOUNDS_ERROR = "'{type}' date {date} is out of bounds"
 UNACCENT_TYPE_ERROR = "argument must be a string; int given"
-FIELD_NONE_ERROR = "'{name}' cannot be None"
-FIELD_EMPTY_ERROR = "'{name}' cannot be an empty string"
-FIELD_WHITESPACES_ERROR = "'{name}' cannot be composed by whitespaces only"
-FIELD_TYPE_ERROR = "field value must be a string; int given"
-
-
-class TestMergeDatetimeRanges(TestCase):
-    """Unit tests for merge_datetime_ranges function"""
-
-    def test_merge_datetime_ranges(self):
-        """Test function with several case scenarios"""
-
-        # Case 1: merge overlapped ranges
-        dates = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2010, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(2008, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-
-        ranges = [r for r in merge_datetime_ranges(dates)]
-        expected = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-        self.assertListEqual(ranges, expected)
-
-        # Case 2: merge a range contained within other ranges
-        dates = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2010, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(2010, 1, 2, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(2008, 1, 1, tzinfo=UTC), datetime.datetime(2010, 1, 1, tzinfo=UTC))
-        ]
-
-        ranges = [r for r in merge_datetime_ranges(dates)]
-        expected = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2010, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(2010, 1, 2, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-        self.assertListEqual(ranges, expected)
-
-        # Case 3: merge a set of ranges contained within another one
-        dates = [
-            (datetime.datetime(2006, 1, 1, tzinfo=UTC), datetime.datetime(2008, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(1999, 1, 1, tzinfo=UTC), datetime.datetime(2000, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-
-        ranges = [r for r in merge_datetime_ranges(dates)]
-        expected = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-        self.assertListEqual(ranges, expected)
-
-        # Case 4: merge disjoint ranges
-        dates = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2010, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(2010, 1, 2, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-
-        ranges = [r for r in merge_datetime_ranges(dates)]
-        expected = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2010, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(2010, 1, 2, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-        self.assertListEqual(ranges, expected)
-
-        # Case 5: merge ranges where part of them are contained within others
-        dates = [
-            (datetime.datetime(2005, 1, 1, tzinfo=UTC), datetime.datetime(2008, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(2005, 10, 15, tzinfo=UTC), datetime.datetime(2010, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(2008, 1, 5, tzinfo=UTC), datetime.datetime(2009, 1, 1, tzinfo=UTC))
-        ]
-
-        ranges = [r for r in merge_datetime_ranges(dates)]
-        expected = [
-            (datetime.datetime(2005, 1, 1, tzinfo=UTC), datetime.datetime(2010, 1, 1, tzinfo=UTC))
-        ]
-        self.assertListEqual(ranges, expected)
-
-        # Case 6: min limit is not excluded but ranges are merged
-        dates = [
-            (datetime.datetime(2000, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-
-        ranges = [r for r in merge_datetime_ranges(dates)]
-        expected = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-        self.assertListEqual(ranges, expected)
-
-        # Case 7: max limit is not excluded but ranges are merged
-        dates = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2000, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-
-        ranges = [r for r in merge_datetime_ranges(dates)]
-        expected = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-        self.assertListEqual(ranges, expected)
-
-        # Case 8: limit ranges are not excluded
-        dates = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-
-        ranges = [r for r in merge_datetime_ranges(dates)]
-        expected = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-        self.assertListEqual(ranges, expected)
-
-        # Case 8a: min limit range is not excluded
-        dates = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2005, 1, 1, tzinfo=UTC))
-        ]
-
-        ranges = [r for r in merge_datetime_ranges(dates)]
-        expected = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2005, 1, 1, tzinfo=UTC))
-        ]
-        self.assertListEqual(ranges, expected)
-
-        # Case 8b: max limit range is not excluded
-        dates = [
-            (datetime.datetime(2005, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-
-        ranges = [r for r in merge_datetime_ranges(dates)]
-        expected = [
-            (datetime.datetime(2005, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-        self.assertListEqual(ranges, expected)
-
-    def test_merge_duplicated_ranges(self):
-        """Test if duplicated ranges are merged into one"""
-
-        dates = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-
-        ranges = [r for r in merge_datetime_ranges(dates)]
-        expected = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-        self.assertListEqual(ranges, expected)
-
-    def test_merge_excluding_limits(self):
-        """Test function with several case inputs when exclude_limits is set"""
-
-        # Case 1: merge overlapped ranges excluding min and max limits
-        dates = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2010, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(2008, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-
-        ranges = [r for r in merge_datetime_ranges(dates, exclude_limits=True)]
-        expected = [
-            (datetime.datetime(2008, 1, 1, tzinfo=UTC), datetime.datetime(2010, 1, 1, tzinfo=UTC))
-        ]
-        self.assertListEqual(ranges, expected)
-
-        # Case 2: merge a range contained within other ranges excluding min and max limits
-        dates = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2010, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(2010, 1, 2, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(2008, 1, 1, tzinfo=UTC), datetime.datetime(2010, 1, 1, tzinfo=UTC))
-        ]
-
-        ranges = [r for r in merge_datetime_ranges(dates, exclude_limits=True)]
-        expected = [
-            (datetime.datetime(2008, 1, 1, tzinfo=UTC), datetime.datetime(2010, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(2010, 1, 2, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-        self.assertListEqual(ranges, expected)
-
-        # Case 3: merge a set of ranges contained within another one excluding min and max limits
-        dates = [
-            (datetime.datetime(2006, 1, 1, tzinfo=UTC), datetime.datetime(2008, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(1999, 1, 1, tzinfo=UTC), datetime.datetime(2000, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-
-        ranges = [r for r in merge_datetime_ranges(dates, exclude_limits=True)]
-        expected = [
-            (datetime.datetime(1999, 1, 1, tzinfo=UTC), datetime.datetime(2000, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(2006, 1, 1, tzinfo=UTC), datetime.datetime(2008, 1, 1, tzinfo=UTC))
-        ]
-        self.assertListEqual(ranges, expected)
-
-        # Case 4: merge disjoint ranges; as they are disjoint, limits are not excluded
-        dates = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2010, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(2010, 1, 2, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-
-        ranges = [r for r in merge_datetime_ranges(dates, exclude_limits=True)]
-        expected = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2010, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(2010, 1, 2, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-        self.assertListEqual(ranges, expected)
-
-        # Case 5: merge ranges where part of them are contained within others
-        dates = [
-            (datetime.datetime(2005, 1, 1, tzinfo=UTC), datetime.datetime(2008, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(2005, 10, 15, tzinfo=UTC), datetime.datetime(2010, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(2008, 1, 5, tzinfo=UTC), datetime.datetime(2009, 1, 1, tzinfo=UTC))
-        ]
-
-        ranges = [r for r in merge_datetime_ranges(dates, exclude_limits=True)]
-        expected = [
-            (datetime.datetime(2005, 1, 1, tzinfo=UTC), datetime.datetime(2010, 1, 1, tzinfo=UTC))
-        ]
-        self.assertListEqual(ranges, expected)
-
-        # Case 6: min limit is excluded and ranges are merged
-        dates = [
-            (datetime.datetime(2000, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-
-        ranges = [r for r in merge_datetime_ranges(dates, exclude_limits=True)]
-        expected = [
-            (datetime.datetime(2000, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-        self.assertListEqual(ranges, expected)
-
-        # Case 7: max limit is excluded and ranges are merged
-        dates = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2000, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-
-        ranges = [r for r in merge_datetime_ranges(dates, exclude_limits=True)]
-        expected = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2000, 1, 1, tzinfo=UTC))
-        ]
-        self.assertListEqual(ranges, expected)
-
-        # Case 8: limit ranges are not excluded
-        dates = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-
-        ranges = [r for r in merge_datetime_ranges(dates, exclude_limits=True)]
-        expected = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-        self.assertListEqual(ranges, expected)
-
-        # Case 8a: min limit range is not excluded
-        dates = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2005, 1, 1, tzinfo=UTC))
-        ]
-
-        ranges = [r for r in merge_datetime_ranges(dates, exclude_limits=True)]
-        expected = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2005, 1, 1, tzinfo=UTC))
-        ]
-        self.assertListEqual(ranges, expected)
-
-        # Case 8b: max limit range is not excluded
-        dates = [
-            (datetime.datetime(2005, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-
-        ranges = [r for r in merge_datetime_ranges(dates, exclude_limits=True)]
-        expected = [
-            (datetime.datetime(2005, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-        self.assertListEqual(ranges, expected)
-
-    def test_merge_duplicated_ranges_excluding_limits(self):
-        """Test if duplicated ranges are merged excluding limits"""
-
-        # On this case, limits will not be excluded due to there is only one range
-        dates = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-
-        ranges = [r for r in merge_datetime_ranges(dates, exclude_limits=True)]
-        expected = [
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC))
-        ]
-        self.assertListEqual(ranges, expected)
-
-    def test_none_list_of_dates(self):
-        """Check if the result is empty when the list of ranges is None"""
-
-        ranges = [r for r in merge_datetime_ranges(None)]
-        self.assertEqual(ranges, [])
-
-    def test_empty_list_of_dates(self):
-        """Check if the result is empty when the list of ranges is empty"""
-
-        ranges = [r for r in merge_datetime_ranges([])]
-        self.assertEqual(ranges, [])
-
-    def test_dates_out_of_bounds(self):
-        """Check whether it raises an exception when dates are out of bounds"""
-
-        # Case 1
-        dates = [
-            (datetime.datetime(2008, 1, 1, tzinfo=UTC), datetime.datetime(2100, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(1800, 1, 1, tzinfo=UTC), datetime.datetime(2010, 1, 1, tzinfo=UTC))
-        ]
-
-        expected = DATE_OUT_OF_BOUNDS_ERROR.format(type='start',
-                                                   date=r'1800-01-01 00:00:00\+00:00')
-        with self.assertRaisesRegexp(ValueError, expected):
-            _ = [r for r in merge_datetime_ranges(dates)]
-
-        # Case 2
-        dates = [
-            (datetime.datetime(2008, 1, 1, tzinfo=UTC), datetime.datetime(2100, 2, 1, tzinfo=UTC)),
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2010, 1, 1, tzinfo=UTC))
-        ]
-
-        expected = DATE_OUT_OF_BOUNDS_ERROR.format(type='end',
-                                                   date=r'2100-02-01 00:00:00\+00:00')
-        with self.assertRaisesRegexp(ValueError, expected):
-            _ = [r for r in merge_datetime_ranges(dates)]
-
-    def test_dates_no_timezone(self):
-        """Check whether it raises an exception when dates without timezone are given"""
-
-        # Case 1
-        dates = [
-            (datetime.datetime(2008, 1, 1), datetime.datetime(2100, 1, 1, tzinfo=UTC)),
-            (datetime.datetime(1800, 1, 1, tzinfo=UTC), datetime.datetime(2010, 1, 1, tzinfo=UTC))
-        ]
-
-        with self.assertRaisesRegexp(TypeError, CANT_COMPARE_DATES_ERROR):
-            _ = [r for r in merge_datetime_ranges(dates)]
-
-        # Case 2
-        dates = [
-            (datetime.datetime(2008, 1, 1, tzinfo=UTC), datetime.datetime(2100, 2, 1, tzinfo=UTC)),
-            (datetime.datetime(1900, 1, 1, tzinfo=UTC), datetime.datetime(2010, 1, 1))
-        ]
-
-        with self.assertRaisesRegexp(TypeError, CANT_COMPARE_DATES_ERROR):
-            _ = [r for r in merge_datetime_ranges(dates)]
+IDENTITY_NONE_OR_EMPTY_ERROR = "identity data cannot be empty"
+SOURCE_NONE_OR_EMPTY_ERROR = "'source' cannot be"
 
 
 class TestUnnacentString(TestCase):
@@ -401,49 +51,97 @@ class TestUnnacentString(TestCase):
             unaccent_string(1234)
 
 
-class TestValidateField(TestCase):
-    """Unit tests for validate_field"""
+class TestUUID(TestCase):
+    """Unit tests for generate_uuid function"""
 
-    def test_validate_string(self):
-        """Check valid fields"""
+    def test_uuid(self):
+        """Check whether the function returns the expected UUID"""
 
-        # If the field is valid, the method does not raise any exception
-        validate_field('test_field', 'Test')
+        result = generate_uuid('scm', email='jsmith@example.com',
+                               name='John Smith', username='jsmith')
+        self.assertEqual(result, 'a9b403e150dd4af8953a52a4bb841051e4b705d9')
 
-    def test_invalid_string(self):
-        """Check invalid string fields"""
+        result = generate_uuid('scm', email='jsmith@example.com')
+        self.assertEqual(result, '334da68fcd3da4e799791f73dfada2afb22648c6')
 
-        expected = FIELD_EMPTY_ERROR.format(name='test_field')
-        with self.assertRaisesRegex(ValueError, expected):
-            validate_field('test_field', '')
+        result = generate_uuid('scm', email='', name='John Smith', username='jsmith')
+        self.assertEqual(result, 'a4b4591c3a2171710c157d7c278ea3cc03becf81')
 
-        expected = FIELD_WHITESPACES_ERROR.format(name='test_field')
-        with self.assertRaisesRegex(ValueError, expected):
-            validate_field('test_field', '\t')
+        result = generate_uuid('scm', email='', name='John Smith', username='')
+        self.assertEqual(result, '76e3624e24aacae178d05352ad9a871dfaf81c13')
 
-        expected = FIELD_WHITESPACES_ERROR.format(name='test_field')
-        with self.assertRaisesRegex(ValueError, expected):
-            validate_field('test_field', '  \t ')
+        result = generate_uuid('scm', email='', name='', username='jsmith')
+        self.assertEqual(result, '6e7ce2426673f8a23a72a343b1382dda84c0078b')
 
-    def test_allow_none(self):
-        """Check valid and invalid fields allowing `None` values"""
+        result = generate_uuid('scm', email='', name='John Ca\xf1as', username='jcanas')
+        self.assertEqual(result, 'c88e126749ff006eb1eea25e4bb4c1c125185ed2')
 
-        validate_field('test_field', None, allow_none=True)
+        result = generate_uuid('scm', email='', name="Max Müster", username='mmuester')
+        self.assertEqual(result, '9a0498297d9f0b7e4baf3e6b3740d22d2257367c')
 
-        expected = FIELD_NONE_ERROR.format(name='test_field')
-        with self.assertRaisesRegex(ValueError, expected):
-            validate_field('test_field', None, allow_none=False)
+    def test_case_insensitive(self):
+        """Check if same values in lower or upper case produce the same UUID"""
 
-        expected = FIELD_EMPTY_ERROR.format(name='test_field')
-        with self.assertRaisesRegex(ValueError, expected):
-            validate_field('test_field', '', allow_none=True)
+        uuid_a = generate_uuid('scm', email='jsmith@example.com',
+                               name='John Smith', username='jsmith')
+        uuid_b = generate_uuid('SCM', email='jsmith@example.com',
+                               name='John Smith', username='jsmith')
+        self.assertEqual(uuid_a, uuid_b)
 
-        expected = FIELD_WHITESPACES_ERROR.format(name='test_field')
-        with self.assertRaisesRegex(ValueError, expected):
-            validate_field('test_field', ' \t ', allow_none=True)
+        uuid_c = generate_uuid('scm', email='jsmith@example.com',
+                               name='john smith', username='jsmith')
+        self.assertEqual(uuid_c, uuid_a)
 
-    def test_no_string(self):
-        """Check if an exception is raised when the value type is not a string"""
+        uuid_d = generate_uuid('scm', email='jsmith@example.com',
+                               name='John Smith', username='JSmith')
+        self.assertEqual(uuid_d, uuid_a)
 
-        with self.assertRaisesRegex(TypeError, FIELD_TYPE_ERROR):
-            validate_field('test_field', 42)
+        uuid_e = generate_uuid('scm', email='JSMITH@example.com',
+                               name='John Smith', username='jsmith')
+        self.assertEqual(uuid_e, uuid_a)
+
+    def test_case_unaccent_name(self):
+        """Check if same values accent or unaccent produce the same UUID"""
+
+        accent_result = generate_uuid('scm', email='', name="Max Müster", username='mmuester')
+        unaccent_result = generate_uuid('scm', email='', name="Max Muster", username='mmuester')
+        self.assertEqual(accent_result, unaccent_result)
+        self.assertEqual(accent_result, '9a0498297d9f0b7e4baf3e6b3740d22d2257367c')
+
+        accent_result = generate_uuid('scm', email='', name="Santiago Dueñas", username='')
+        unaccent_result = generate_uuid('scm', email='', name="Santiago Duenas", username='')
+        self.assertEqual(accent_result, unaccent_result)
+        self.assertEqual(accent_result, '0f1dd18839007ee8a11d02572ca0a0f4eedaf2cd')
+
+        accent_result = generate_uuid('scm', email='', name="Tomáš Čechvala", username='')
+        partial_accent_result = generate_uuid('scm', email='', name="Tomáš Cechvala", username='')
+        unaccent_result = generate_uuid('scm', email='', name="Tomas Cechvala", username='')
+        self.assertEqual(accent_result, unaccent_result)
+        self.assertEqual(accent_result, partial_accent_result)
+
+    def test_surrogate_escape(self):
+        """Check if no errors are raised for invalid UTF-8 chars"""
+
+        result = generate_uuid('scm', name="Mishal\udcc5 Pytasz")
+        self.assertEqual(result, '625166bdc2c4f1a207d39eb8d25315010babd73b')
+
+    def test_none_source(self):
+        """Check whether UUID cannot be obtained giving a None source"""
+
+        with self.assertRaisesRegex(ValueError, SOURCE_NONE_OR_EMPTY_ERROR):
+            generate_uuid(None)
+
+    def test_empty_source(self):
+        """Check whether UUID cannot be obtained giving aadded to the registry"""
+
+        with self.assertRaisesRegex(ValueError, SOURCE_NONE_OR_EMPTY_ERROR):
+            generate_uuid('')
+
+    def test_none_or_empty_data(self):
+        """Check whether UUID cannot be obtained when identity data is None or empty"""
+
+        with self.assertRaisesRegex(ValueError, IDENTITY_NONE_OR_EMPTY_ERROR):
+            generate_uuid('scm', email=None, name='', username=None)
+
+        with self.assertRaisesRegex(ValueError, IDENTITY_NONE_OR_EMPTY_ERROR):
+            generate_uuid('scm', email='', name='', username='')
